Source Code/requirements.txt
absl-py==0.1.11
altgraph==0.15
astor==0.6.2
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
Cython==0.28.3
future==0.16.0
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
Keras==2.1.5
kiwisolver==1.0.1
lxml==4.1.1
macholib==1.9
Markdown==2.6.11
matplotlib==2.2.0
numpy==1.14.1
opencv-python==3.4.0.12
panda==0.3.1
pandas==0.22.0
pandas-datareader==0.6.0
pefile==2017.11.5
Pillow==5.0.0
protobuf==3.5.2
PyInstaller==3.3.1
pyparsing==2.2.0
pypiwin32==223
python-dateutil==2.6.1
pytz==2018.3
pywin32==223
PyYAML==3.12
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
